Move: user 패키지 domain 하위로 이관
