+++ src/main/java/com/babjo/deliverycommerce/domain/user/dto/AdminSignupResponseDto.java
@@ -1,13 +1,12 @@
-package com.babjo.deliverycommerce.user.dto;
+package com.babjo.deliverycommerce.domain.user.dto;
 
-import com.babjo.deliverycommerce.user.entity.User;
+import com.babjo.deliverycommerce.domain.user.entity.User;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class UserResponseDto {
-
+public class AdminSignupResponseDto {
     private final Long userId;
     private final String username;
     private final String nickname;
@@ -16,7 +15,7 @@ public class UserResponseDto {
     private final LocalDateTime createdAt;
 
     // 엔티티 -> DTO 변환은 DTO 생성자 쪽에서 처리
-    public UserResponseDto(User user) {
+    public AdminSignupResponseDto(User user) {
         this.userId = user.getUserId();
         this.username = user.getUsername();
         this.nickname = user.getNickname();
