+++ src/main/java/com/babjo/deliverycommerce/domain/user/dto/AdminUserUpdateRoleResponseDto.java
@@ -1,4 +1,4 @@
-package com.babjo.deliverycommerce.user.dto;
+package com.babjo.deliverycommerce.domain.user.dto;
 
 import lombok.Getter;
