+++ src/main/java/com/babjo/deliverycommerce/domain/user/dto/UserResponseDto.java
@@ -1,28 +1,27 @@
-package com.babjo.deliverycommerce.user.dto;
+package com.babjo.deliverycommerce.domain.user.dto;
 
-import com.babjo.deliverycommerce.user.entity.User;
+import com.babjo.deliverycommerce.domain.user.entity.User;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
 
 @Getter
-public class UserSearchResponseDto {
+public class UserResponseDto {
+
     private final Long userId;
     private final String username;
     private final String nickname;
     private final String email;
     private final String role;
     private final LocalDateTime createdAt;
-    private final LocalDateTime deletedAt;  // 목록에서만 노출
 
-    public UserSearchResponseDto(User user) {
+    // 엔티티 -> DTO 변환은 DTO 생성자 쪽에서 처리
+    public UserResponseDto(User user) {
         this.userId = user.getUserId();
         this.username = user.getUsername();
         this.nickname = user.getNickname();
         this.email = user.getEmail();
         this.role = user.getRole().name();
         this.createdAt = user.getCreatedAt();
-        this.deletedAt = user.getDeletedAt();
-
     }
 }
